lambda/requirements.txt
boto3>=1.26.0
strands-agents>=1.4.0
